+++ spider/app/models/data_models/__init__.py
@@ -1,10 +1,10 @@
+""" This module contains domain models used and returned by core components and services.
+"""
+
 from .data_models import (
     DataModel,
-    JobResult,
-    HTMLData,
-    URL,
-    RequestHeader,
-    ParseRule,
-    ParseResult,
-    CrawlResult
+    RequestHeader
 )
+from .spider_models import HTMLData, URL
+from .parser_models import ParseResult
+from .crawler_models import CrawlResult
